Add files via upload
This is the final submission and all my changes made to me website. this included cleaning up any mistakes and making sure everything was cohesive and matched. Making changes to the pages to make sure there were no errors such as headers being the same colour and also adding in some images