Add files via upload
this is my changes so far to my website to match the HTML/A2 canva layout. Will be changing the HTML files to suit me more as well as changing the images to ones i need and match my orginal canva layout. 
had to have it on seperate pages as unable to figure out a scroll feature over multiples pages
still trying to figure out the night/day toggle